sqlite migration for simplicity

+++ backend/models.py
@@ -1,30 +1,30 @@
-from sqlalchemy import Column, BigInteger, Integer, String, Float, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, String, Float, Boolean, ForeignKey
 from database import Base
 
 
 class Player(Base):
     __tablename__ = "players"
 
-    id = Column(BigInteger, primary_key=True)  # OpenDota account_id
+    id = Column(Integer, primary_key=True)  # OpenDota account_id
     name = Column(String)
 
 
 class Match(Base):
     __tablename__ = "matches"
 
-    match_id = Column(BigInteger, primary_key=True)
-    radiant_team_id = Column(BigInteger)
-    dire_team_id = Column(BigInteger)
-    league_id = Column(BigInteger, ForeignKey("leagues.id"))
+    match_id = Column(Integer, primary_key=True)
+    radiant_team_id = Column(Integer)
+    dire_team_id = Column(Integer)
+    league_id = Column(Integer, ForeignKey("leagues.id"))
 
 
 class PlayerMatchStats(Base):
     __tablename__ = "player_match_stats"
 
-    id = Column(BigInteger, primary_key=True)
-    player_id = Column(BigInteger, ForeignKey("players.id"))
-    match_id = Column(BigInteger, ForeignKey("matches.match_id"))
-    team_id = Column(BigInteger, ForeignKey("teams.id"))
+    id = Column(Integer, primary_key=True)
+    player_id = Column(Integer, ForeignKey("players.id"))
+    match_id = Column(Integer, ForeignKey("matches.match_id"))
+    team_id = Column(Integer, ForeignKey("teams.id"))
     fantasy_points = Column(Float)
 
     # Raw stats stored so fantasy points can be recalculated without re-fetching
@@ -40,25 +40,25 @@ class PlayerMatchStats(Base):
 class Team(Base):
     __tablename__ = "teams"
 
-    id = Column(BigInteger, primary_key=True)  # OpenDota team_id
+    id = Column(Integer, primary_key=True)  # OpenDota team_id
     name = Column(String)
 
 
 class Card(Base):
     __tablename__ = "cards"
 
-    id = Column(BigInteger, primary_key=True)
-    player_id = Column(BigInteger, ForeignKey("players.id"))
-    owner_id = Column(BigInteger, ForeignKey("users.id"))
+    id = Column(Integer, primary_key=True)
+    player_id = Column(Integer, ForeignKey("players.id"))
+    owner_id = Column(Integer, ForeignKey("users.id"))
     card_type = Column(String)  # "common", "rare", "epic", "legendary"
-    league_id = Column(BigInteger, ForeignKey("leagues.id"))
+    league_id = Column(Integer, ForeignKey("leagues.id"))
     is_active = Column(Boolean, default=False)
 
 
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     username = Column(String, unique=True)
     email = Column(String, unique=True)
     password_hash = Column(String)
@@ -68,7 +68,7 @@ class User(Base):
 class League(Base):
     __tablename__ = "leagues"
 
-    id = Column(BigInteger, primary_key=True)  # OpenDota league_id
+    id = Column(Integer, primary_key=True)  # OpenDota league_id
     name = Column(String)
